create a distinct component to display the form reservation and the calendar
